starter code refactor
* Top level statement
* nuget update
* remove json.net
* file scoped namespaces
* implicit usings
* connection string from configuration
* rename project and solution
