+++ Bme.DataDriven.Rest.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32526.322
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "rest", "rest\rest.csproj", "{9CEE4BF1-EBCA-476D-9CC2-88F3F4088246}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bme.DataDriven.Rest", "Bme.DataDriven.Rest\Bme.DataDriven.Rest.csproj", "{9CEE4BF1-EBCA-476D-9CC2-88F3F4088246}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
