+++ Bme.DataDriven.Rest/Dal/DataDrivenDbContext.cs
@@ -1,0 +1,265 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace Bme.DataDriven.Rest.Dal;
+
+public partial class DataDrivenDbContext : DbContext
+{
+    public DataDrivenDbContext(DbContextOptions<DataDrivenDbContext> options)
+        : base(options)
+    {
+    }
+
+    public virtual DbSet<Category> Category { get; set; }
+    public virtual DbSet<Customer> Customer { get; set; }
+    public virtual DbSet<CustomerSite> CustomerSite { get; set; }
+    public virtual DbSet<Invoice> Invoice { get; set; }
+    public virtual DbSet<InvoiceIssuer> InvoiceIssuer { get; set; }
+    public virtual DbSet<InvoiceItem> InvoiceItem { get; set; }
+    public virtual DbSet<Order> Order { get; set; }
+    public virtual DbSet<OrderItem> OrderItem { get; set; }
+    public virtual DbSet<PaymentMethod> PaymentMethod { get; set; }
+    public virtual DbSet<Product> Product { get; set; }
+    public virtual DbSet<Status> Status { get; set; }
+    public virtual DbSet<VAT> Vat { get; set; }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<Category>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.Name).HasMaxLength(50);
+
+            entity.Property(e => e.ParentCategoryId).HasColumnName("ParentCategoryID");
+
+            entity.HasOne(d => d.ParentCategory)
+                .WithMany(p => p.InverseParentCategory)
+                .HasForeignKey(d => d.ParentCategoryId);
+        });
+
+        modelBuilder.Entity<Customer>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.BankAccount)
+                .HasMaxLength(50)
+                .IsUnicode(false);
+
+            entity.Property(e => e.Email)
+                .HasMaxLength(50)
+                .IsUnicode(false);
+
+            entity.Property(e => e.Login).HasMaxLength(50);
+
+            entity.Property(e => e.MainCustomerSiteId).HasColumnName("MainCustomerSiteID");
+
+            entity.Property(e => e.Name).HasMaxLength(50);
+
+            entity.Property(e => e.Password).HasMaxLength(50);
+
+            entity.HasOne(d => d.MainCustomerSite)
+                .WithMany(p => p.Customer)
+                .HasForeignKey(d => d.MainCustomerSiteId);
+        });
+
+        modelBuilder.Entity<CustomerSite>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.City).HasMaxLength(50);
+
+            entity.Property(e => e.CustomerId).HasColumnName("CustomerID");
+
+            entity.Property(e => e.Fax)
+                .HasMaxLength(15)
+                .IsUnicode(false);
+
+            entity.Property(e => e.Street).HasMaxLength(50);
+
+            entity.Property(e => e.Tel)
+                .HasMaxLength(15)
+                .IsUnicode(false);
+
+            entity.Property(e => e.ZipCode)
+                .HasMaxLength(4)
+                .IsUnicode(false);
+
+            entity.HasOne(d => d.MainCustomer)
+                .WithMany(p => p.CustomerSites)
+                .HasForeignKey(d => d.CustomerId);
+        });
+
+        modelBuilder.Entity<Invoice>(entity =>
+        {
+            entity.Property(e => e.Id)
+                .HasColumnName("ID")
+                .ValueGeneratedNever();
+
+            entity.Property(e => e.CreationDate).HasColumnType("datetime");
+
+            entity.Property(e => e.CustomerCity).HasMaxLength(50);
+
+            entity.Property(e => e.CustomerName).HasMaxLength(50);
+
+            entity.Property(e => e.CustomerStreet).HasMaxLength(50);
+
+            entity.Property(e => e.CustomerZipCode)
+                .HasMaxLength(4)
+                .IsUnicode(false);
+
+            entity.Property(e => e.DeliveryDate).HasColumnType("datetime");
+
+            entity.Property(e => e.InvoiceIssuerId).HasColumnName("InvoiceIssuerID");
+
+            entity.Property(e => e.OrderId).HasColumnName("OrderID");
+
+            entity.Property(e => e.PaymentDeadline).HasColumnType("datetime");
+
+            entity.Property(e => e.PaymentMethod).HasMaxLength(20);
+
+            entity.HasOne(d => d.InvoiceIssuer)
+                .WithMany(p => p.Invoices)
+                .HasForeignKey(d => d.InvoiceIssuerId);
+
+            entity.HasOne(d => d.Order)
+                .WithMany(p => p.Invoice)
+                .HasForeignKey(d => d.OrderId);
+        });
+
+        modelBuilder.Entity<InvoiceIssuer>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.BankAccount)
+                .HasMaxLength(50)
+                .IsUnicode(false);
+
+            entity.Property(e => e.City).HasMaxLength(50);
+
+            entity.Property(e => e.Name).HasMaxLength(50);
+
+            entity.Property(e => e.Street).HasMaxLength(50);
+
+            entity.Property(e => e.TaxIdentifier)
+                .HasMaxLength(20)
+                .IsUnicode(false);
+
+            entity.Property(e => e.ZipCode)
+                .HasMaxLength(4)
+                .IsUnicode(false);
+        });
+
+        modelBuilder.Entity<InvoiceItem>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.InvoiceId).HasColumnName("InvoiceID");
+
+            entity.Property(e => e.Name).HasMaxLength(50);
+
+            entity.Property(e => e.OrderItemId).HasColumnName("OrderItemID");
+
+            entity.Property(e => e.Vatpercentage).HasColumnName("VATPercentage");
+
+            entity.HasOne(d => d.Invoice)
+                .WithMany(p => p.InvoiceItems)
+                .HasForeignKey(d => d.InvoiceId);
+
+            entity.HasOne(d => d.OrderItem)
+                .WithMany(p => p.InvoiceItems)
+                .HasForeignKey(d => d.OrderItemId);
+        });
+
+        modelBuilder.Entity<Order>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.CustomerSiteId).HasColumnName("CustomerSiteID");
+
+            entity.Property(e => e.Date).HasColumnType("datetime");
+
+            entity.Property(e => e.Deadline).HasColumnType("datetime");
+
+            entity.Property(e => e.PaymentMethodId).HasColumnName("PaymentMethodID");
+
+            entity.Property(e => e.StatusId).HasColumnName("StatusID");
+
+            entity.HasOne(d => d.CustomerSite)
+                .WithMany(p => p.Orders)
+                .HasForeignKey(d => d.CustomerSiteId);
+
+            entity.HasOne(d => d.PaymentMethod)
+                .WithMany(p => p.Orders)
+                .HasForeignKey(d => d.PaymentMethodId);
+
+            entity.HasOne(d => d.Status)
+                .WithMany(p => p.Orders)
+                .HasForeignKey(d => d.StatusId);
+        });
+
+        modelBuilder.Entity<OrderItem>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.OrderId).HasColumnName("OrderID");
+
+            entity.Property(e => e.ProductId).HasColumnName("ProductID");
+
+            entity.Property(e => e.StatusId).HasColumnName("StatusID");
+
+            entity.HasOne(d => d.Order)
+                .WithMany(p => p.OrderItems)
+                .HasForeignKey(d => d.OrderId);
+
+            entity.HasOne(d => d.Product)
+                .WithMany(p => p.OrderItem)
+                .HasForeignKey(d => d.ProductId);
+
+            entity.HasOne(d => d.Status)
+                .WithMany(p => p.OrderItems)
+                .HasForeignKey(d => d.StatusId);
+        });
+
+        modelBuilder.Entity<PaymentMethod>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.Method).HasMaxLength(20);
+        });
+
+        modelBuilder.Entity<Product>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.CategoryId).HasColumnName("CategoryID");
+
+            entity.Property(e => e.Description).HasColumnType("xml");
+
+            entity.Property(e => e.Name).HasMaxLength(50);
+
+            entity.Property(e => e.VatId).HasColumnName("VATID");
+
+            entity.HasOne(d => d.Category)
+                .WithMany(p => p.Products)
+                .HasForeignKey(d => d.CategoryId);
+
+            entity.HasOne(d => d.VAT)
+                .WithMany(p => p.Product)
+                .HasForeignKey(d => d.VatId);
+        });
+
+        modelBuilder.Entity<Status>(entity =>
+        {
+            entity.Property(e => e.Id).HasColumnName("ID");
+
+            entity.Property(e => e.Name).HasMaxLength(20);
+        });
+
+        modelBuilder.Entity<VAT>(entity =>
+        {
+            entity.ToTable("VAT");
+
+            entity.Property(e => e.Id).HasColumnName("ID");
+        });
+    }
+}
