+++ Bme.DataDriven.Rest/Dal/Status.cs
@@ -1,0 +1,16 @@
+﻿namespace Bme.DataDriven.Rest.Dal;
+
+public partial class Status
+{
+    public Status()
+    {
+        Orders = new HashSet<Order>();
+        OrderItems = new HashSet<OrderItem>();
+    }
+
+    public int Id { get; set; }
+    public string Name { get; set; }
+
+    public ICollection<Order> Orders { get; set; }
+    public ICollection<OrderItem> OrderItems { get; set; }
+}
