+++ Bme.DataDriven.Rest/Properties/launchSettings.json
@@ -1,15 +1,7 @@
 {
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:57264",
-      "sslPort": 0
-    }
-  },
   "$schema": "http://json.schemastore.org/launchsettings.json",
   "profiles": {
-    "rest": {
+    "Bme.DataDriven.Rest": {
       "commandName": "Project",
       "launchUrl": "api/values",
       "environmentVariables": {
