+++ Bme.DataDriven.Rest/appsettings.Development.json
